Code Cleanup

+++ app/src/main/java/com/example/stromtracker/ui/urlaub/urlaub_new/UrlaubNewFragment.kt
@@ -80,7 +80,7 @@ class UrlaubNewFragment(private val geraete: List<Geraete>, private val currHaus
                 if (geraet.getStromStandBy() != null && geraet.getBetriebszeitStandBy() != null)
                     gesamtverbrauchNeuPT += geraet.getStromStandBy()!! * dayLen
                 else {
-                    gesamtverbrauchNeuPT += geraet.getStromVollast() * geraet.getBetriebszeit() as Double
+                    gesamtverbrauchNeuPT += geraet.getStromVollast() * geraet.getBetriebszeit()
                 }
             }
             gesamtverbrauchAktPT += geraet.getJahresverbrauch()
